+++ puppetboard/forms.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 
 from flask.ext.wtf import Form
 from wtforms import (
-    HiddenField, RadioField, SelectField, 
+    HiddenField, RadioField, SelectField,
     TextAreaField, validators
 )
 
@@ -17,9 +17,14 @@ class QueryForm(Form):
         ('nodes', 'Nodes'),
         ('resources', 'Resources'),
         ('facts', 'Facts'),
-        ('fact-names', 'Fact Names'),
+        ('factsets', 'Fact Sets'),
+        ('fact-paths', 'Fact Paths'),
+        ('fact-contents', 'Fact Contents'),
         ('reports', 'Reports'),
         ('events', 'Events'),
+        ('catalogs', 'Catalogs'),
+        ('edges', 'Edges'),
+        ('environments', 'Environments'),
         ])
 
 class CatalogForm(Form):