+++ puppetboard/utils.py
@@ -3,6 +3,7 @@ from __future__ import unicode_literals
 
 import json
 
+from math import ceil
 from requests.exceptions import HTTPError, ConnectionError
 from pypuppetdb.errors import EmptyResponseError
 
@@ -29,17 +30,6 @@ def get_or_abort(func, *args, **kwargs):
         abort(204)
 
 
-def limit_reports(reports, limit):
-    """Helper to yield a number of from the reports generator.
-
-    This is an ugly solution at best...
-    """
-    for count, report in enumerate(reports):
-        if count == limit:
-            raise StopIteration
-        yield report
-
-
 def yield_or_stop(generator):
     """Similar in intent to get_or_abort this helper will iterate over our
     generators and handle certain errors.
@@ -54,3 +44,35 @@ def yield_or_stop(generator):
             raise
         except (EmptyResponseError, ConnectionError, HTTPError):
             raise StopIteration
+
+class Pagination(object):
+
+    def __init__(self, page, per_page, total_count):
+        self.page = page
+        self.per_page = per_page
+        self.total_count = total_count
+
+    @property
+    def pages(self):
+        return int(ceil(self.total_count / float(self.per_page)))
+
+    @property
+    def has_prev(self):
+        return self.page > 1
+
+    @property
+    def has_next(self):
+        return self.page < self.pages
+
+    def iter_pages(self, left_edge=2, left_current=2,
+                   right_current=5, right_edge=2):
+        last = 0
+        for num in xrange(1, self.pages + 1):
+            if num <= left_edge or \
+               (num > self.page - left_current - 1 and \
+                num < self.page + right_current) or \
+                num > self.pages - right_edge:
+                if last + 1 != num:
+                    yield None
+                yield num
+                last = num